Default request interceptor
An interceptor that is able to provide default values for any part of the
standard request api:
- method
- path
- params
- headers
- entity

If a value is not defined in the request, this interceptor will provide the
configured default value. For headers and params, the values are mixed
together. This interceptor will not overwrite a value specfied by a request.