Indicate the originating client for a request
The client that originated the request will not be included on the
request object as the 'originator'.

The 'originator' property may be inacurate if an interceptor replaces the
request object without copying this property forward.

This commit also, finally, included unit tests for rest/interceptor.  The
behavior for interceptors has also been clarified for request handlers
that return a rejected promise; the response is rejected with the value
of the promise.

Issue: #24 #10