RFC 5988 compatible Link header parsing
The hateoas interceptor now parses Link response headers indexing them
under response.links. Multiple Link headers are supported.

The value of the Link header must match the grammar defined in rfc5988. If
the grammar is incompatible, the link is skipped.

http://tools.ietf.org/html/rfc5988

PEG.js (http://pegjs.majda.cz/) is used to generate the parser for the
rfc defined grammer. The generated parser is wrapped in UMD and checked
into the repo to avoid a run/build-time dependency on PEG.js.