JSON Hypertext Application Language
Adding mime converter support for application/hal+json. Objects are read as
HAL indexing links and embedded objects on to the resource. Objects are
written as plain JSON.

Embedded relationships are indexed onto the resource by the relationship as
a promise for the related resource.

Links are indexed onto the resource as a lazy promise that will GET the
resource when a handler is first registered on the promise.

A `clientFor` method is added to the entity to get a full Client for a
relationship.

The `_links` and `_embedded` properties on the resource are made
non-enumerable.

Issue #37