Disable XHR cancellation test
This test is misbehaving in IE 6, the response is received before we have
an opportunity to cancel the request.  In order to properly test this
component, we need an HTTP endpoint that is slow to respond. Buster 0.7
will enable this capability.

Disabling this test until we have a proper solution.