First pass at streams in rest.js [Experimental]
- Support steams in all clients and most interceptors
- Pull in Node Streams API via browserify for browsers
- Drop support for Node 0.6 and 0.8 (the Streams API is deficient)

Interceptors have the ability to request a streamed request/response
entity instead of a fully buffered entity. The previous interceptor life
cycle hooks work as is, new '_____Stream' hooks provide the same life
cycle phases, but with a stream entity. The interceptor model will
automatically convert to/from streams as required. As the end consumer of
an interceptor chain, it's currently ambiguous whether a stream or
concrete entity will be contained in the response.