HTTP Response Promise helper methods
Basic methods included on the HTTP response promise to make interacting
with portions of the HTTP reasponse easier.

    var response = client(request);

    response.entity().then(function (entity) { ... });
    response.status().then(function (statusCode) { ... });
    response.headers().then(function (headers) { ... });
    response.header('Content-Type').then(function (contentType) { ... });

Of course all the typical when.js promise goodness is also available on
the promise.

The resolution status of promises returned from these methods mirors the
resolution state of the response promise: fulfillment->fulfillemnt,
rejection->rejection.

If a non-traditional response object is used to resolve the promise,
these method may behave strangly. In most cases, they will resolve with
undefined, but the behavior itself is not defined.