Allow custom JSON revivers and replacers
The JSON converter may be extended to provide a custom reviver and/or
replacer. The custom converter must then be registered into a MIME
registry before it can be consumed by the MIME interceptor.

This CL also introduced delegate converters within a registry. A delegate
is a dynamically bound converter that proxies the read and write methods
to the target.

The '+json' converter now delegates to the 'application/json' converter.

Fixes: #50