Cancel requests via the response promise
The ResponsePromise now contains a `cancel` method. The method will
attempt to cancel the request is there is a cancel method available on
the request, otherwise, it will set the canceled property on the request
which will prevent a request from being initiated.

    var responsePromise = client(request);
    responsePromise.cancel();
    responsePromise.catch(function (canceledResponse) {
        ...
    });

For the cancel method to work, the client making the request must support
cancelation. All built in clients support cancelation, but third-party
clients may not.

Issue: #118