Overwrites config params with request params
Before this commit the config params passed in when wrapping the client
with the template interceptor could not be overwritten by the request
params passed in when invoking the client.

By changing the order of the arguments in the interceptor/template.js
mixin function, the config params (defualt params) can be overwritten by
the request params.

Also in this commit:

- Update year in licence block

- Add test to confirm config params overwrite behavior

Issue: #180